feat(config): add extraRoutes to scully.config (#139)
* feat(config): add extraRoutes to scully.config

With the additional of extraRoutes to the scully.config, users can now provide additional routes
that they want Scully to discover. These routes can plug into the router plugins as well, and can
produce multitudes of browseable routes each.

* improvement(scully): promisify the extraRoutes process

I converted this process to return a `Promise<unhandledRoutes[]>` which is what we needed it to be
in order to allow for async fetching of extraRoutes at build time. This is actually a very wonderful
and robust method.