fix(remove httpclientmodule): refactor to drop depedency on httpClien… (#182)
* fix(remove httpclientmodule): refactor to drop depedency on httpClientModule

This refactor takes care of the httpClientModule, by replacing it with a fetchHttp from our own
making.

* feat(transferstate): refactor to use a global, instead of parsion on intial page

update the tranferstate to prepare it for external JSON, and a little optimized flow

* improvement(transer-state.service): use index.html to prevent unneeded 301

read the index.html instead of the urls itself. Saves a round-trip to the server.

* fix(transferstate): make sure there are no emmisions of state _during_ routing

Make sure state is only emitted after routing has finished, to make sure the state doesn't get
transfered to the old page.

* improvement(transerstate): make sure state doesn't get emmitted during navigation

This makes sure the state doesn't get emmited before navigation is done.

* fix(transferstate): move navigationDome to correct place

* improvement(transferstateservice): don't emit state during navigation

with this, the getstate will _never_ emit a value **during** navigation. Only when navigation is
done, the state will be available.