fix(scully): Add wait time before spawning new browser instance (#804)
If there are more than 500 routes rendered, when this limit is reached
a new browser instance is spawned. The async jobs rendering the
routes at that moment in the initial browser instance are not waited
and thus they are failing with exception and cause the process to stop.
Update and add testserver data for 500+ routes.
Update test server to hold and serve 501 users.
Update snapshots for the newly added users
Add test to check extra routes.