feat(scully): scully-plugin-flash-prevention Hide app-root with position absolute
There are systems which rely on their components to take up space in order to
lay them out
correctly. For example, angular material with an open side nav - the content (mat-sidenav-content)
is given margin-left of 200px
but the js calculations responsible for that dont work if it's
display:none, so its not added when I use scully.
As a workaround, I added the margin in css, but
that's not ideal. A better fix is to have the app-root take up space in the DOM

BREAKING CHANGE:
Currently, I switch body:not(.loaded) app-root to be position absolute (and covered by
app-root-scully) instead of display: none. Its possible that that may cause problems.