docs(ng-lib): add useScullyTransferState description (#1246)
* docs(transfer-state): add method coverage, add resolver info

Added description for `useScullyTransferState`. Additionally provided some information regarding the usage of TransferState with Angular Route Resolvers.

* chore(ng-lib): undo addition about resolver

After further testing in my app, I could not make it work with this approach, altough it initially seemed to work. Will have to further investigate.. For now, it should not be part of the docs..