fix(ng-lib): scully-content only upgrades href attributes that are projected inside it.
scully-content only modifies hrefs that are projected into it. Previously scully-content was
modifying ALL app links to be Angular router friendly. This may be okay in some cases, but not all,
sometimes we just want an href to behave outside Angular router. The fact that using a
<scully-content> can modify the behaviour of hrefs not contained inside scully-content is strange
and unexpected. This change fixes this.